update kieu du lieu cua ma tran

+++ src/main/java/SumValueDiagonalLineOfMatrix.java
@@ -5,8 +5,8 @@ public class SumValueDiagonalLineOfMatrix {
         Scanner scanner = new Scanner(System.in);
         System.out.println("Nhap vao kich thuoc cua ma tran");
         int size = scanner.nextInt();
-        int[][] matrix = new int[size][size];
-        int sum = 0;
+        double[][] matrix = new double[size][size];
+        double sum = 0;
 
         for (int i = 0; i < size; i++) {
             for (int j = 0; j < size; j++) {
